DiscreteUpdateManager computes contact kinematics
The computation of DiscreteContactPair data and ContactKinematics data
is moved from CompliantContactManager to the base class
DiscreteUpdateManager. This is in line with computation of
ContactResults and ContactSolverResults. This allows objects with only
a pointer to the base class to access these information.
